Moved some more code out of the main file, and re-added some cors and rate-limiting middleware which was lost.
